feat(storage,registry): modularize cloud storage starters and support split container images
