--- registry/src/main/java/io/terrakube/registry/plugin/storage/aws/AwsStorageServiceProperties.java
@@ -1,22 +1,0 @@
-package io.terrakube.registry.plugin.storage.aws;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
-
-@Component
-@Getter
-@Setter
-@PropertySource(value = "classpath:application.properties", ignoreResourceNotFound = true)
-@PropertySource(value = "classpath:application-${spring.profiles.active}.properties", ignoreResourceNotFound = true)
-@ConfigurationProperties(prefix = "io.terrakube.registry.plugin.storage.aws")
-public class AwsStorageServiceProperties {
-    private String accessKey;
-    private String secretKey;
-    private String bucketName;
-    private String region;
-    private String endpoint;
-    private boolean enableRoleAuthentication;
-}
