--- registry/src/main/java/io/terrakube/registry/plugin/storage/azure/AzureStorageServiceProperties.java
@@ -1,19 +1,0 @@
-package io.terrakube.registry.plugin.storage.azure;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
-
-@Component
-@Getter
-@Setter
-@PropertySource(value = "classpath:application.properties", ignoreResourceNotFound = true)
-@PropertySource(value = "classpath:application-${spring.profiles.active}.properties", ignoreResourceNotFound = true)
-@ConfigurationProperties(prefix = "io.terrakube.registry.plugin.storage.azure")
-public class AzureStorageServiceProperties {
-
-    private String accountName;
-    private String accountKey;
-}
