--- registry/src/main/java/io/terrakube/registry/service/git/GitService.java
@@ -1,8 +1,0 @@
-package io.terrakube.registry.service.git;
-
-import java.io.File;
-
-public interface GitService {
-
-    File getCloneRepositoryByTag(ModuleVersionDownload download);
-}
