+++ registry/src/main/java/io/terrakube/registry/service/module/ModuleServiceImpl.java
@@ -10,8 +10,8 @@ import io.terrakube.client.model.organization.module.ModuleRequest;
 import io.terrakube.client.model.organization.ssh.Ssh;
 import io.terrakube.client.model.organization.vcs.Vcs;
 import io.terrakube.client.model.organization.vcs.github_app_token.GitHubAppToken;
-import io.terrakube.registry.plugin.storage.StorageService;
-import io.terrakube.registry.service.git.ModuleVersionDownload;
+import io.terrakube.storage.plugin.core.StorageService;
+import io.terrakube.storage.plugin.core.ModuleVersionDownload;
 import io.terrakube.registry.service.search.CommonSearchService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
