+++ storage/storage-azure/src/main/java/io/terrakube/storage/plugin/azure/AzureStorageServiceProperties.java
@@ -1,0 +1,15 @@
+package io.terrakube.storage.plugin.azure;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "io.terrakube.registry.plugin.storage.azure")
+public class AzureStorageServiceProperties {
+    private String accountName;
+    private String accountKey;
+}
