+++ storage/storage-core/src/main/java/io/terrakube/storage/plugin/core/GitService.java
@@ -1,0 +1,8 @@
+package io.terrakube.storage.plugin.core;
+
+import java.io.File;
+
+public interface GitService {
+
+    File getCloneRepositoryByTag(ModuleVersionDownload download);
+}
