+++ storage/storage-core/src/main/java/io/terrakube/storage/plugin/core/StorageType.java
@@ -1,0 +1,8 @@
+package io.terrakube.storage.plugin.core;
+
+public enum StorageType {
+    Local,
+    AzureStorageImpl,
+    AwsStorageImpl,
+    GcpStorageImpl
+}
